+++ ps_shcolar/py/styles.py
@@ -1,22 +1,11 @@
 __doc__="""
 This module contains escape codes to set (reset) styles of following text
 """
+from . import lists
+from . import consts
 
-reset="\033[49m";
-""" Reset all colors/styles to default"""
-bold="\033[1m";
-"""Escape sequence for bold text"""
-dim="\033[2m";
-"""Escape sequence for dimmed text"""
-italic="\033[3m";
-"""Escape sequence for italic text"""
-
-underline="\033[4m";
-"""Escape sequence for underlined text"""
-
-invert="\033[7m";
-"""Escape sequence for inverted text"""
-
-underline="\033[4m";
-"""Escape sequence for underlined text"""
-
+pfx_str=consts.sgr_prefix_str
+sfx_str=consts.sgr_suffix_str
+styles={itm["name"]:pfx_str+itm["on"]+sfx_str for itm in lists.styles_list}
+styles_off={itm["name"]:pfx_str+itm["off"]+sfx_str for itm in lists.styles_list}
+reset='\033[0m'