+++ ps_funcs/py/CHANGELOG.md
@@ -139,3 +139,7 @@ there in main module
 - fixed indebtarion errors
 ### [files.py](files.py)
 - addded `start of` message
+
+
+### [__init__.py](__init__.py)
+- changed `ps_files` back to `files` again